swapped long and short

+++ data.js
@@ -1,12 +1,12 @@
 module.exports = {
-  short:
+  long:
   [
     { title: 'It\'s time for your scheduled break', body: 'Take a break, have a KitKat' },
     { title: 'It\'s time for your scheduled break', body: 'Taking time to do nothing often brings everything into perspective...' },
     { title: 'It\'s time for your scheduled break', body: 'Grab yourself a cup of tea!' },
     { title: 'It\'s time for your scheduled break', body: '%REPLACE% minutes for meditation!' },
   ],
-  long:
+  short:
   [
     { title: 'It\'s time to look away', body: 'Quick, your %REPLACE% seconds starts now...' },
     { title: 'It\'s time to look away', body: 'See anything interesting 20ft away?' },
